Move OrderedHash to Mapi module to prevent conflicts with other implementations.
Specifically, OrderedHash can conflict with the OrderedHash defined in
sass 2.4.7 running under ruby 1.8.7
